better db scripts

+++ database_backup.sh
@@ -12,29 +12,33 @@ REPO="$(basename "$DIR")"
 BACKUPS_DIR="$DIR/backups"
 FILENAME=$REPO-`date +%s` # "date +%s" returns the epoch timestamp
 
-# Import the database host, username, and password
+# Import the database information
 source "$DIR/.env"
 if [[ -z $DB_HOST ]]; then
   DB_HOST=localhost
+fi
+if [[ -z $DB_PORT ]]; then
+  DB_PORT=5432
 fi
 
 # Back up the database
 # (we specify an extension of ".bak", which means that pg_dump will automatically compress the data)
 mkdir -p "$BACKUPS_DIR"
-echo Dumping the database...
-PGPASSWORD="$DB_PASS" pg_dump --host="$DB_HOST" --username="$DB_USER" "$DB_NAME" > "$BACKUPS_DIR/$FILENAME.bak"
+echo "Dumping the database..."
+PGPASSWORD="$DB_PASS" pg_dump --host="$DB_HOST" --port="$DB_PORT" --username="$DB_USER" --dbname="$DB_NAME" > "$BACKUPS_DIR/$FILENAME.bak"
 if [[ $? -ne 0 ]]; then
   exit 1
 fi
-#echo Zipping the backup...
-#gzip "$BACKUPS_DIR/$FILENAME"
-echo Complete.
+echo "Dump complete."
 
 # Delete old backups if the hard drive is getting full
-AMOUNT_FULL=$(df "$DIR" | tail -1 | awk '{print $5}' | rev | cut -c 2- | rev)
-if [[ $AMOUNT_FULL -gt 80 ]]; then
-  # Delete the oldest file in the backups directory
-  rm "$(ls -t "$BACKUPS_DIR" | tail -1)"
+# (but skip this if we are on Windows)
+if uname -a | grep -v MINGW64 >/dev/null 2>&1; then
+  AMOUNT_FULL=$(df "$DIR" | tail -1 | awk '{print $5}' | rev | cut -c 2- | rev)
+  if [[ $AMOUNT_FULL -gt 80 ]]; then
+    # Delete the oldest file in the backups directory
+    rm "$(ls -t "$BACKUPS_DIR" | tail -1)"
+  fi
 fi
 
 # Upload it to Google Drive
@@ -49,6 +53,7 @@ fi
 if command -v gdrive > /dev/null; then
   if [[ ! -z $GOOGLE_DRIVE_SERVICE_ACCOUNT ]]; then
     gdrive upload "$BACKUPS_DIR/$FILENAME" --service-account "$GOOGLE_DRIVE_SERVICE_ACCOUNT" --parent "$GOOGLE_DRIVE_PARENT_DIRECTORY"
+    echo "Upload complete."
   else
     echo "Skipping upload to Google Drive since \"GOOGLE_DRIVE_SERVICE_ACCOUNT\" is not set in the \".env\" file."
   fi
